+++ airflow/pipe_vessels_dag.py
@@ -1,11 +1,14 @@
-from datetime import datetime, timedelta
-
 from airflow import DAG
-from airflow.operators.bash_operator import BashOperator
 from airflow.models import Variable
+from airflow.models import Variable
+from airflow.operators.bash_operator import BashOperator
 
 from airflow_ext.gfw import config as config_tools
 from airflow_ext.gfw.models import DagFactory
+from airflow_ext.gfw.operators.helper.flexible_operator import FlexibleOperator
+
+from datetime import datetime, timedelta
+
 
 PIPELINE = "pipe_vessels"
 
@@ -19,53 +22,71 @@ class VesselsPipelineDagFactory(DagFactory):
         return '{}.{}'.format(prefix, vessel_pipe_name)
 
     def build(self, dag_id):
-        
+
         config = self.config
         config.update(self.vessel_config)
 
         with DAG(dag_id, schedule_interval=self.schedule_interval, default_args=self.default_args) as dag:
-           
+
             config['date_range'] = ','.join(self.source_date_range())
             name = config['name']
+            flexible_operator_var = Variable.get('FLEXIBLE_OPERATOR')
 
-            publish_vessel_info = BashOperator(
-                task_id='publish_vessel_info_{}'.format(name),
-                depends_on_past=True,
-                bash_command='{docker_run} {docker_image} publish_vessel_info '
-                '\'{bigquery_vessel_info_query}\' '
-                '{project_id}:{temp_dataset} '
-                '{temp_bucket} '
-                '{elasticsearch_server_url} '
-                '{elasticsearch_server_auth} '
-                '{elasticsearch_index_alias} '
-                '\'{elasticsearch_index_mappings}\''.format(**config)
-            )
+            publish_vessel_info_params = {
+                'task_id':'publish_vessel_info_{}'.format(name),
+                'depends_on_past':True,
+                'pool':'k8operators_limit' if flexible_operator_var == 'kubernetes' else 'bigquery',
+                'docker_run':'{docker_run}'.format(**config),
+                'image':'{docker_image}'.format(**config),
+                'name':'pipe-vessels-publish-vessel-info-{}'.format(name),
+                'dag':dag,
+                'arguments':['publish_vessel_info',
+                             '{bigquery_vessel_info_query}'.format(**config),
+                             '{project_id}:{temp_dataset}'.format(**config),
+                             '{temp_bucket}'.format(**config),
+                             '{elasticsearch_server_url}'.format(**config),
+                             '{elasticsearch_server_auth}'.format(**config),
+                             '{elasticsearch_index_alias}'.format(**config),
+                             '{elasticsearch_index_mappings}'.format(**config)]
+            }
+            publish_vessel_info = FlexibleOperator(publish_vessel_info_params).build_operator(flexible_operator_var)
 
-            aggregate_tracks = BashOperator(
-                task_id='aggregate_tracks_{}'.format(name),
-                pool='bigquery',
-                depends_on_past=True,
-                bash_command='{docker_run} {docker_image} aggregate_tracks '
-                '{date_range} '
-                '\'{bigquery_vessel_tracks_jinja_query}\' '
-                '{project_id}:{target_dataset}.{bigquery_tracks} '.format(
-                    **config)
-            )
+            aggregate_tracks_params = {
+                'task_id':'aggregate_tracks_{}'.format(name),
+                'depends_on_past':True,
+                'pool':'k8operators_limit' if flexible_operator_var == 'kubernetes' else 'bigquery',
+                'docker_run':'{docker_run}'.format(**config),
+                'image':'{docker_image}'.format(**config),
+                'name':'pipe-vessels-aggregate-tracks-{}'.format(name),
+                'dag':dag,
+                'arguments':['aggregate_tracks',
+                             '{date_range}'.format(**config),
+                             '{bigquery_vessel_tracks_jinja_query}'.format(**config),
+                             '{project_id}:{target_dataset}.{bigquery_tracks}'.format(**config)]
+            }
+            aggregate_tracks = FlexibleOperator(aggregate_tracks_params).build_operator(flexible_operator_var)
 
-            publish_postgres_tracks = BashOperator(
-                task_id='publish_postgres_tracks_{}'.format(name),
-                depends_on_past=True,
-                bash_command='{docker_run} {docker_image} publish_postgres_tracks '
-                '{date_range} '
-                '{project_id}:{target_dataset}.{bigquery_tracks} '
-                '{project_id}:{temp_dataset} '
-                '{temp_bucket} '
-                '{postgres_instance} '
-                '{postgres_connection_string} '
-                '{postgres_table_tracks}'.format(**config)
-            )
+            publish_postgres_tracks_params = {
+                'task_id':'publish_postgres_tracks_{}'.format(name),
+                'depends_on_past':True,
+                'pool':'k8operators_limit' if flexible_operator_var == 'kubernetes' else 'bigquery',
+                'docker_run':'{docker_run}'.format(**config),
+                'image':'{docker_image}'.format(**config),
+                'name':'pipe-vessels-publish-postgres-tracks-{}'.format(name),
+                'dag':dag,
+                'arguments':['publish_postgres_tracks',
+                             '{date_range}'.format(**config),
+                             '{project_id}:{target_dataset}.{bigquery_tracks}'.format(**config),
+                             '{project_id}:{temp_dataset}'.format(**config),
+                             '{temp_bucket}'.format(**config),
+                             '{postgres_instance}'.format(**config),
+                             '{postgres_connection_string}'.format(**config),
+                             '{postgres_table_tracks}'.format(**config)]
+            }
+            publish_postgres_tracks = FlexibleOperator(publish_postgres_tracks_params).build_operator(flexible_operator_var)
 
-            check_source_existance = config.get('check_source_existance',None)
+
+            check_source_existance = config.get('check_source_existance', None)
             if (check_source_existance is not None or not check_source_existance):
                 dag >> aggregate_tracks
                 dag >> publish_vessel_info
@@ -86,7 +107,7 @@ for mode in modes:
     for vessels_configuration in vessels_configurations:
         dag_factory = VesselsPipelineDagFactory(vessels_configuration,schedule_interval='@{}'.format(mode))
         dag_id = dag_factory.get_dag_id(
-                            '{}_{}'.format(PIPELINE, mode),
-                            vessels_configuration['name']
-                        )
+            '{}_{}'.format(PIPELINE, mode),
+            vessels_configuration['name']
+        )
         globals()[dag_id] = dag_factory.build(dag_id=dag_id)